+++ src/app/shared/BibleBooks.ts
@@ -1,5 +1,0 @@
-export interface BibleBooks {
-    bookName: string,
-    bookId: number,
-    subcategory: string,
-}