+++ src/app/shared/bible/Book.ts
@@ -1,0 +1,4 @@
+export interface Book {
+    bookName: string, 
+    bookId: number,
+}