+++ src/app/shared/bible/Chapter.ts
@@ -1,0 +1,4 @@
+export interface Chapter {
+    chapterName: string, 
+    chapterId: number,
+}