+++ src/app/shared/bible/Verse.ts
@@ -1,0 +1,4 @@
+export interface Verse {
+    verseName: string, 
+    verseId: number,
+}